tests/test_utils.py
# tests/test_utils.py

import os
import pytest
from src.utils import long_path, validate_directory_structure

# Mocking the os.path.exists method for testing
def mock_exists(path):
    if path == "C:\\Scans\\pending":
        return True  # Simulate that this directory exists
    if path == "C:\\Photos-repo\\pending":
        return True  # Simulate that this directory exists
    return False  # Simulate that other directories do not exist

def test_long_path_windows():
    """Test long path handling on Windows."""
    if os.name == 'nt':  # Check if the OS is Windows
        assert long_path("C:\\Some\\Path") == "\\\\?\\C:\\Some\\Path"
    else:
        assert long_path("C:\\Some\\Path") == "C:\\Some\\Path"

def test_validate_directory_structure(monkeypatch):
    """Test directory structure validation."""
    # Mocking os.path.exists to control the behavior of the function being tested
    monkeypatch.setattr(os.path, "exists", mock_exists)
    
    # Assuming DESTINATION_BASE is set correctly in your settings or mock it as needed
    pdf_source = "C:\\Scans\\pending"
    photo_source_base = "C:\\Photos-repo\\pending"
    destination_base = "C:\\Users\\hgavilanes\\KEMCO Industries\\KEMCO - Documents"
    
    assert validate_directory_structure(pdf_source, photo_source_base, destination_base) is True

# Add more tests as needed for other functions...